requirements.txt
# KSR-Net: Low-Resolution License Plate Recognition
# ICPR 2026 Competition

# ============================================
# Core Deep Learning
# ============================================
torch>=2.0.0
torchvision>=0.15.0

# ============================================
# Image Processing
# ============================================
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
scikit-image>=0.21.0

# ============================================
# FlowFormer++ Dependencies
# ============================================
timm==0.5.4          # Specific version for FlowFormer++ compatibility
einops>=0.7.0
yacs>=0.1.8
loguru>=0.7.0

# ============================================
# Text Recognition (MGP-STR)
# ============================================
transformers>=4.30.0
sentencepiece>=0.1.99

# ============================================
# Utilities
# ============================================
tqdm>=4.65.0
matplotlib>=3.7.0

# ============================================
# Development & Testing
# ============================================
pytest>=7.0.0
